+++ library/include/internal/hipsolver-compat.h
@@ -20,6 +20,8 @@
  ********************************************************************************/
 typedef hipsolverHandle_t hipsolverDnHandle_t;
 
+typedef void* hipsolverGesvdjInfo_t;
+
 typedef void* hipsolverSyevjInfo_t;
 
 #ifdef __cplusplus
@@ -65,6 +67,28 @@ HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXsyevjGetResidual(hipsolverDnHandl
 HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXsyevjGetSweeps(hipsolverDnHandle_t  handle,
                                                               hipsolverSyevjInfo_t info,
                                                               int*                 executed_sweeps);
+
+// params
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnCreateGesvdjInfo(hipsolverGesvdjInfo_t* info);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnDestroyGesvdjInfo(hipsolverGesvdjInfo_t info);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXgesvdjSetMaxSweeps(hipsolverGesvdjInfo_t info,
+                                                                  int                   max_sweeps);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXgesvdjSetSortEig(hipsolverGesvdjInfo_t info,
+                                                                int                   sort_eig);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXgesvdjSetTolerance(hipsolverGesvdjInfo_t info,
+                                                                  double                tolerance);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXgesvdjGetResidual(hipsolverDnHandle_t   handle,
+                                                                 hipsolverGesvdjInfo_t info,
+                                                                 double*               residual);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnXgesvdjGetSweeps(hipsolverDnHandle_t   handle,
+                                                               hipsolverGesvdjInfo_t info,
+                                                               int* executed_sweeps);
 
 // orgbr/ungbr
 HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnSorgbr_bufferSize(hipsolverHandle_t   handle,
@@ -963,6 +987,268 @@ HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnZgesvd(hipsolverHandle_t handle,
                                                      int               lwork,
                                                      double*           rwork,
                                                      int*              devInfo);
+
+// gesvdj
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnSgesvdj_bufferSize(hipsolverDnHandle_t   handle,
+                                                                 hipsolverEigMode_t    jobz,
+                                                                 int                   econ,
+                                                                 int                   m,
+                                                                 int                   n,
+                                                                 float*                A,
+                                                                 int                   lda,
+                                                                 float*                S,
+                                                                 float*                U,
+                                                                 int                   ldu,
+                                                                 float*                V,
+                                                                 int                   ldv,
+                                                                 int*                  lwork,
+                                                                 hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnDgesvdj_bufferSize(hipsolverDnHandle_t   handle,
+                                                                 hipsolverEigMode_t    jobz,
+                                                                 int                   econ,
+                                                                 int                   m,
+                                                                 int                   n,
+                                                                 double*               A,
+                                                                 int                   lda,
+                                                                 double*               S,
+                                                                 double*               U,
+                                                                 int                   ldu,
+                                                                 double*               V,
+                                                                 int                   ldv,
+                                                                 int*                  lwork,
+                                                                 hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnCgesvdj_bufferSize(hipsolverDnHandle_t   handle,
+                                                                 hipsolverEigMode_t    jobz,
+                                                                 int                   econ,
+                                                                 int                   m,
+                                                                 int                   n,
+                                                                 hipFloatComplex*      A,
+                                                                 int                   lda,
+                                                                 float*                S,
+                                                                 hipFloatComplex*      U,
+                                                                 int                   ldu,
+                                                                 hipFloatComplex*      V,
+                                                                 int                   ldv,
+                                                                 int*                  lwork,
+                                                                 hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnZgesvdj_bufferSize(hipsolverDnHandle_t   handle,
+                                                                 hipsolverEigMode_t    jobz,
+                                                                 int                   econ,
+                                                                 int                   m,
+                                                                 int                   n,
+                                                                 hipDoubleComplex*     A,
+                                                                 int                   lda,
+                                                                 double*               S,
+                                                                 hipDoubleComplex*     U,
+                                                                 int                   ldu,
+                                                                 hipDoubleComplex*     V,
+                                                                 int                   ldv,
+                                                                 int*                  lwork,
+                                                                 hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnSgesvdj(hipsolverDnHandle_t   handle,
+                                                      hipsolverEigMode_t    jobz,
+                                                      int                   econ,
+                                                      int                   m,
+                                                      int                   n,
+                                                      float*                A,
+                                                      int                   lda,
+                                                      float*                S,
+                                                      float*                U,
+                                                      int                   ldu,
+                                                      float*                V,
+                                                      int                   ldv,
+                                                      float*                work,
+                                                      int                   lwork,
+                                                      int*                  devInfo,
+                                                      hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnDgesvdj(hipsolverDnHandle_t   handle,
+                                                      hipsolverEigMode_t    jobz,
+                                                      int                   econ,
+                                                      int                   m,
+                                                      int                   n,
+                                                      double*               A,
+                                                      int                   lda,
+                                                      double*               S,
+                                                      double*               U,
+                                                      int                   ldu,
+                                                      double*               V,
+                                                      int                   ldv,
+                                                      double*               work,
+                                                      int                   lwork,
+                                                      int*                  devInfo,
+                                                      hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnCgesvdj(hipsolverDnHandle_t   handle,
+                                                      hipsolverEigMode_t    jobz,
+                                                      int                   econ,
+                                                      int                   m,
+                                                      int                   n,
+                                                      hipFloatComplex*      A,
+                                                      int                   lda,
+                                                      float*                S,
+                                                      hipFloatComplex*      U,
+                                                      int                   ldu,
+                                                      hipFloatComplex*      V,
+                                                      int                   ldv,
+                                                      hipFloatComplex*      work,
+                                                      int                   lwork,
+                                                      int*                  devInfo,
+                                                      hipsolverGesvdjInfo_t params);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnZgesvdj(hipsolverDnHandle_t   handle,
+                                                      hipsolverEigMode_t    jobz,
+                                                      int                   econ,
+                                                      int                   m,
+                                                      int                   n,
+                                                      hipDoubleComplex*     A,
+                                                      int                   lda,
+                                                      double*               S,
+                                                      hipDoubleComplex*     U,
+                                                      int                   ldu,
+                                                      hipDoubleComplex*     V,
+                                                      int                   ldv,
+                                                      hipDoubleComplex*     work,
+                                                      int                   lwork,
+                                                      int*                  devInfo,
+                                                      hipsolverGesvdjInfo_t params);
+
+// gesvdj_batched
+HIPSOLVER_EXPORT hipsolverStatus_t
+    hipsolverDnSgesvdjBatched_bufferSize(hipsolverDnHandle_t   handle,
+                                         hipsolverEigMode_t    jobz,
+                                         int                   m,
+                                         int                   n,
+                                         float*                A,
+                                         int                   lda,
+                                         float*                S,
+                                         float*                U,
+                                         int                   ldu,
+                                         float*                V,
+                                         int                   ldv,
+                                         int*                  lwork,
+                                         hipsolverGesvdjInfo_t params,
+                                         int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t
+    hipsolverDnDgesvdjBatched_bufferSize(hipsolverDnHandle_t   handle,
+                                         hipsolverEigMode_t    jobz,
+                                         int                   m,
+                                         int                   n,
+                                         double*               A,
+                                         int                   lda,
+                                         double*               S,
+                                         double*               U,
+                                         int                   ldu,
+                                         double*               V,
+                                         int                   ldv,
+                                         int*                  lwork,
+                                         hipsolverGesvdjInfo_t params,
+                                         int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t
+    hipsolverDnCgesvdjBatched_bufferSize(hipsolverDnHandle_t   handle,
+                                         hipsolverEigMode_t    jobz,
+                                         int                   m,
+                                         int                   n,
+                                         hipFloatComplex*      A,
+                                         int                   lda,
+                                         float*                S,
+                                         hipFloatComplex*      U,
+                                         int                   ldu,
+                                         hipFloatComplex*      V,
+                                         int                   ldv,
+                                         int*                  lwork,
+                                         hipsolverGesvdjInfo_t params,
+                                         int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t
+    hipsolverDnZgesvdjBatched_bufferSize(hipsolverDnHandle_t   handle,
+                                         hipsolverEigMode_t    jobz,
+                                         int                   m,
+                                         int                   n,
+                                         hipDoubleComplex*     A,
+                                         int                   lda,
+                                         double*               S,
+                                         hipDoubleComplex*     U,
+                                         int                   ldu,
+                                         hipDoubleComplex*     V,
+                                         int                   ldv,
+                                         int*                  lwork,
+                                         hipsolverGesvdjInfo_t params,
+                                         int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnSgesvdjBatched(hipsolverDnHandle_t   handle,
+                                                             hipsolverEigMode_t    jobz,
+                                                             int                   m,
+                                                             int                   n,
+                                                             float*                A,
+                                                             int                   lda,
+                                                             float*                S,
+                                                             float*                U,
+                                                             int                   ldu,
+                                                             float*                V,
+                                                             int                   ldv,
+                                                             float*                work,
+                                                             int                   lwork,
+                                                             int*                  devInfo,
+                                                             hipsolverGesvdjInfo_t params,
+                                                             int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnDgesvdjBatched(hipsolverDnHandle_t   handle,
+                                                             hipsolverEigMode_t    jobz,
+                                                             int                   m,
+                                                             int                   n,
+                                                             double*               A,
+                                                             int                   lda,
+                                                             double*               S,
+                                                             double*               U,
+                                                             int                   ldu,
+                                                             double*               V,
+                                                             int                   ldv,
+                                                             double*               work,
+                                                             int                   lwork,
+                                                             int*                  devInfo,
+                                                             hipsolverGesvdjInfo_t params,
+                                                             int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnCgesvdjBatched(hipsolverDnHandle_t   handle,
+                                                             hipsolverEigMode_t    jobz,
+                                                             int                   m,
+                                                             int                   n,
+                                                             hipFloatComplex*      A,
+                                                             int                   lda,
+                                                             float*                S,
+                                                             hipFloatComplex*      U,
+                                                             int                   ldu,
+                                                             hipFloatComplex*      V,
+                                                             int                   ldv,
+                                                             hipFloatComplex*      work,
+                                                             int                   lwork,
+                                                             int*                  devInfo,
+                                                             hipsolverGesvdjInfo_t params,
+                                                             int                   batch_count);
+
+HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnZgesvdjBatched(hipsolverDnHandle_t   handle,
+                                                             hipsolverEigMode_t    jobz,
+                                                             int                   m,
+                                                             int                   n,
+                                                             hipDoubleComplex*     A,
+                                                             int                   lda,
+                                                             double*               S,
+                                                             hipDoubleComplex*     U,
+                                                             int                   ldu,
+                                                             hipDoubleComplex*     V,
+                                                             int                   ldv,
+                                                             hipDoubleComplex*     work,
+                                                             int                   lwork,
+                                                             int*                  devInfo,
+                                                             hipsolverGesvdjInfo_t params,
+                                                             int                   batch_count);
 
 // getrf
 HIPSOLVER_EXPORT hipsolverStatus_t hipsolverDnSgetrf_bufferSize(